backend/requirements.txt
fastapi 
uvicorn 
llama-cpp-python
pillow 
python-multipart
